refactor: :truck: Move TagList to own file
so that i can use it in LocalPlayer
